+++ R/F_swSite.R
@@ -67,7 +67,9 @@ setClass(
     TranspirationRegions = "matrix",
     swrc_flags = "character",
     has_swrcp = "logical",
-    depth_sapric = "numeric"
+    depth_sapric = "numeric",
+    PotSoilEvCoMethod = "integer",
+    RootingProfileMethod = "integer"
   ),
   prototype = list(
     SWClimits = c(swc_min = NA_real_, swc_init = NA_real_, swc_wet = NA_real_),
@@ -114,7 +116,9 @@ setClass(
     ),
     swrc_flags = c(swrc_name = NA_character_, ptf_name = NA_character_),
     has_swrcp = NA,
-    depth_sapric = NA_real_
+    depth_sapric = NA_real_,
+    PotSoilEvCoMethod = NA_integer_,
+    RootingProfileMethod = NA_integer_
   )
 )
 
@@ -198,6 +202,16 @@ setValidity(
 
     if (length(object@depth_sapric) != 1L) {
       msg <- "@depth_sapric length != 1."
+      val <- if (isTRUE(val)) msg else c(val, msg)
+    }
+
+    if (length(object@PotSoilEvCoMethod) != 1L) {
+      msg <- "@PotSoilEvCoMethod length != 1."
+      val <- if (isTRUE(val)) msg else c(val, msg)
+    }
+
+    if (length(object@RootingProfileMethod) != 1L) {
+      msg <- "@RootingProfileMethod length != 1."
       val <- if (isTRUE(val)) msg else c(val, msg)
     }
 
@@ -405,6 +419,22 @@ setMethod(
   "swSite_SoilDensityInputType",
   "swSite",
   function(object) slot(object, "SoilDensityInputType")
+)
+
+#' @rdname swSite-class
+#' @export
+setMethod(
+  "swSite_PotSoilEvCoMethod",
+  "swSite",
+  function(object) slot(object, "PotSoilEvCoMethod")
+)
+
+#' @rdname swSite-class
+#' @export
+setMethod(
+  "swSite_RootingProfileMethod",
+  "swSite",
+  function(object) slot(object, "RootingProfileMethod")
 )
 
 #' @rdname swSite-class
@@ -631,6 +661,30 @@ setReplaceMethod(
       dim = dim(value),
       dimnames = list(NULL, colnames(object@TranspirationRegions))
     )
+    validObject(object)
+    object
+  }
+)
+
+#' @rdname swSite-class
+#' @export
+setReplaceMethod(
+  "swSite_PotSoilEvCoMethod",
+  signature = "swSite",
+  definition = function(object, value) {
+    object@PotSoilEvCoMethod <- as.integer(value)
+    validObject(object)
+    object
+  }
+)
+
+#' @rdname swSite-class
+#' @export
+setReplaceMethod(
+  "swSite_RootingProfileMethod",
+  signature = "swSite",
+  definition = function(object, value) {
+    object@RootingProfileMethod <- as.integer(value)
     validObject(object)
     object
   }