+++ R/K_swContainer.R
@@ -1713,6 +1713,22 @@ setMethod(
 #' @rdname swInputData-class
 #' @export
 setMethod(
+  "swSite_PotSoilEvCoMethod",
+  signature = "swInputData",
+  function(object) swSite_PotSoilEvCoMethod(object@site)
+)
+
+#' @rdname swInputData-class
+#' @export
+setMethod(
+  "swSite_RootingProfileMethod",
+  signature = "swInputData",
+  function(object) swSite_RootingProfileMethod(object@site)
+)
+
+#' @rdname swInputData-class
+#' @export
+setMethod(
   "swSite_TranspirationRegions",
   signature = "swInputData",
   function(object) swSite_TranspirationRegions(object@site)
@@ -1920,6 +1936,28 @@ setReplaceMethod(
   signature = "swInputData",
   function(object, value) {
     swSite_SoilDensityInputType(object@site) <- value
+    object
+  }
+)
+
+#' @rdname swInputData-class
+#' @export
+setReplaceMethod(
+  "swSite_PotSoilEvCoMethod",
+  signature = "swInputData",
+  function(object, value) {
+    swSite_PotSoilEvCoMethod(object@site) <- value
+    object
+  }
+)
+
+#' @rdname swInputData-class
+#' @export
+setReplaceMethod(
+  "swSite_RootingProfileMethod",
+  signature = "swInputData",
+  function(object, value) {
+    swSite_RootingProfileMethod(object@site) <- value
     object
   }
 )