+++ src/rSW_Site.c
@@ -53,7 +53,9 @@ static char *cSW_SIT[] = {
   "SoilDensityInputType",
   "TranspirationRegions",
   "swrc_flags", "has_swrcp",
-  "depth_sapric"
+  "depth_sapric",
+  "PotSoilEvCoMethod",
+  "RootingProfileMethod"
 };
 
 static char *cLayers[] = {
@@ -392,6 +394,10 @@ SEXP onGet_SW_SIT(void) {
 
 	SEXP SoilDensityInputType;
 
+	SEXP PotSoilEvCoMethod;
+
+	SEXP RootingProfileMethod;
+
 	SEXP TranspirationRegions, TranspirationRegions_names, TranspirationRegions_names_y;
 	char *cTranspirationRegions[] = { "ndx", "layer" };
 	int *p_transp; // ideally `LyrIndex` so that same type as `_TranspRgnBounds`, but R API INTEGER() is signed
@@ -501,6 +507,10 @@ SEXP onGet_SW_SIT(void) {
 
 	PROTECT(SoilDensityInputType = ScalarInteger(si->type_soilDensityInput));
 
+	PROTECT(PotSoilEvCoMethod = ScalarInteger(si->methodEvCo));
+
+	PROTECT(RootingProfileMethod = ScalarInteger(si->methodTrCo));
+
 	PROTECT(TranspirationRegions = allocMatrix(INTSXP, si->n_transp_rgn, 2));
 	p_transp = INTEGER(TranspirationRegions);
 	for (i = 0; i < si->n_transp_rgn; i++) {
@@ -549,8 +559,10 @@ SEXP onGet_SW_SIT(void) {
 	SET_SLOT(SW_SIT, install(cSW_SIT[14]), swrc_flags);
 	SET_SLOT(SW_SIT, install(cSW_SIT[15]), has_swrcp);
 	SET_SLOT(SW_SIT, install(cSW_SIT[16]), depthSapric);
+	SET_SLOT(SW_SIT, install(cSW_SIT[17]), PotSoilEvCoMethod);
+	SET_SLOT(SW_SIT, install(cSW_SIT[18]), RootingProfileMethod);
 
-	UNPROTECT(31);
+	UNPROTECT(33);
 	return SW_SIT;
 }
 
@@ -573,6 +585,8 @@ void onSet_SW_SIT(SEXP SW_SIT, LOG_INFO* LogInfo) {
 	SEXP SoilDensityInputType;
 	SEXP swrc_flags, has_swrcp;
     SEXP depthSapric;
+    SEXP PotSoilEvCoMethod;
+    SEXP RootingProfileMethod;
 
     int unprotects = 0;
 
@@ -709,6 +723,14 @@ void onSet_SW_SIT(SEXP SW_SIT, LOG_INFO* LogInfo) {
     PROTECT(depthSapric = GET_SLOT(SW_SIT, install("depth_sapric")));
 	unprotects++;
     si->depthSapric = REAL(depthSapric)[0];
+
+    PROTECT(PotSoilEvCoMethod = GET_SLOT(SW_SIT, install("PotSoilEvCoMethod")));
+    unprotects++;
+    si->methodEvCo = INTEGER(PotSoilEvCoMethod)[0];
+
+    PROTECT(RootingProfileMethod = GET_SLOT(SW_SIT, install("RootingProfileMethod")));
+    unprotects++;
+    si->methodTrCo = INTEGER(RootingProfileMethod)[0];
 
 freeMem:
     UNPROTECT(unprotects);